combined/requirements.txt
flask
python-dotenv
azure-cognitiveservices-vision-computervision
requests
azure-ai-translation-text==1.0.0b1
azure-cognitiveservices-vision-face
msrest

